% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/addvar.R
\name{addvar}
\alias{addvar}
\alias{addvar<-}
\title{Add variable to (model) object}
\usage{
addvar(x, ...)
}
\arguments{
\item{x}{Model object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for adding variables to model object
}
\author{
Klaus K. Holst
}

