% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/baptize.R
\name{baptize}
\alias{baptize}
\title{Label elements of object}
\usage{
baptize(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for labeling elements of an object
}
\author{
Klaus K. Holst
}

