% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blockdiag}
\alias{blockdiag}
\title{Combine matrices to block diagonal structure}
\usage{
blockdiag(x, ..., pad = 0)
}
\arguments{
\item{x}{Matrix}

\item{pad}{Vyalue outside block-diagonal}

\item{\dots}{Additional matrices}
}
\description{
Combine matrices to block diagonal structure
}
\examples{
A <- diag(3)+1
blockdiag(A,A,A,pad=NA)
}
\author{
Klaus K. Holst
}

