\docType{data}
\name{bmd}
\alias{bmd}
\title{Longitudinal Bone Mineral Density Data (Wide format)}
\format{data.frame}
\source{
  Vonesh & Chinchilli (1997), Table 5.4.1 on page 228.
}
\description{
  Bone Mineral Density Data consisting of 112 girls
  randomized to receive calcium og placebo. Longitudinal
  measurements of bone mineral density (g/cm^2) measured
  approximately every 6th month in 3 years.
}
\seealso{
  calcium
}
\keyword{datasets}

