% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_calculus.R
\name{c.estimate}
\alias{c.estimate}
\title{Concatenate estimate objects}
\usage{
\method{c}{estimate}(..., as.list = FALSE)
}
\arguments{
\item{...}{\code{estimate} objects and/or named numeric values}

\item{as.list}{if TRUE the returned object will be of class \code{list} and not
and \code{estimate} (or \code{estimate.extra}) object.}
}
\value{
An \code{estimate} object (if all args are estimates) or an
\code{estimate.extra} object containing \verb{$estimate} and \verb{$extra} components.
}
\description{
When all arguments are \code{estimate} objects, they are merged into a single
\code{estimate}. When some arguments are not \code{estimate} objects but are named
numeric scalars/vectors, an object of class \code{estimate.extra} is returned.
This is useful for bundling auxiliary per-iteration information (e.g.,
convergence status) alongside an estimate for use with \code{\link[=sim.default]{sim.default()}}.
}
\details{
arguments \code{drop.ic}, \code{paired}, \code{sep} are passed to \link{merge.estimate}
}
\examples{
e <- estimate(coef = c(a = 1, b = 2), vcov = diag(2) * 0.1)
# Bundle estimate with extra information
c(e, converged = 1, niter = 10)
}
\seealso{
\code{\link[=sim.default]{sim.default()}} \code{\link[=merge.estimate]{merge.estimate()}}
}
