% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estimate.R
\name{c.summary.estimate}
\alias{c.summary.estimate}
\title{Concatenate summary.estimate objects}
\usage{
\method{c}{summary.estimate}(...)
}
\arguments{
\item{...}{\code{summary.estimate} objects and/or named numeric values}
}
\description{
When all arguments are \code{summary.estimate} objects, they are merged into a
single \code{summary.estimate} object. When some arguments are not
\code{summary.estimate} objects but are named numeric scalars/vectors, an object
of class \code{summary.estimate} is returned with an additional \code{extra} attribute
that contains the provided numeric scalars/vectors. This is useful for
bundling auxiliary per-iteration information (e.g., convergence status)
alongside an estimate for use with \code{\link[=sim.default]{sim.default()}}.
}
\examples{
e1 <- estimate(coef = 1, IC = scale(rnorm(10)), id = 1:10, labels = "a1")
e2 <- estimate(coef = 2, IC = scale(rnorm(10)), id = 1:10, labels = "a2")

# concatenating two summary.estimate objects
c(e1, e2)

# concatenating one summary.estimate object with one numerical variable
ss <- c(summary(e1), niter = 2)
print(ss)
attributes(ss)$extra
}
