\name{calcium}
\docType{data}
\alias{calcium}
\title{Longitudinal Bone Mineral Density Data}
\description{
  Bone Mineral Density Data consisting of 112 girls randomized
  to receive calcium og placebo. Longitudinal measurements of
  bone mineral density (g/cm^2) measured approximately every 6th 
  month in 3 years.
}
\usage{calcium}
\format{A data.frame containing 112 observations.}
\source{book, p.xxx}
\references{
  ???
}
\keyword{datasets}
