% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/cancel.R
\name{cancel}
\alias{cancel}
\alias{cancel<-}
\title{Generic cancel method}
\usage{
cancel(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Additioal arguments}
}
\description{
Generic cancel method
}
\author{
Klaus K. Holst
}

