% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/children.R
\name{children}
\alias{children}
\alias{parents}
\title{Extract children or parent elements of object}
\usage{
children(object, ...)
}
\arguments{
\item{object}{Object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for extracting children or parent elements of object (e.g. a graph)
}
\author{
Klaus K. Holst
}

