% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commutation.R
\name{commutation}
\alias{commutation}
\title{Finds the unique commutation matrix}
\usage{
commutation(m, n = m)
}
\arguments{
\item{m}{rows}

\item{n}{columns}
}
\description{
Finds the unique commutation matrix K:
\eqn{K vec(A) = vec(A^t)}
}
\author{
Klaus K. Holst
}

