% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Generic method for extracting correlation coefficients of model object}
\usage{
correlation(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Additional arguments}
}
\description{
Generic correlation method
}
\author{
Klaus K. Holst
}

