\name{curereg}
\alias{curereg}
\title{Regression model for binomial data with unkown group of immortals}
\usage{
  curereg(formula, cureformula = ~1, data,
    fam = binomial(), offset = NULL, start, var = "robust",
    ...)
}
\arguments{
  \item{formula}{Formula specifying}

  \item{cureformula}{}

  \item{data}{}

  \item{fam}{}

  \item{offset}{}

  \item{start}{}

  \item{var}{}

  \item{...}{}
}
\description{
  Regression model for binomial data with unkown group of
  immortals
}
\author{
  Klaus K. Holst
}

