% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Cross-validation}
\usage{
cv(modelList, data, K = 5, rep = 1, perf, seed = NULL, mc.cores = 1,
  ...)
}
\arguments{
\item{modelList}{List of fitting functions}

\item{data}{data.frame}

\item{K}{Number of folds (default 5)}

\item{rep}{Number of repetitions (default 1)}

\item{perf}{Performance measure (default RMSE)}

\item{seed}{Optional random seed}

\item{mc.cores}{Number of cores used for parallel computations}

\item{...}{Additional arguments parsed to models in modelList and perf}
}
\description{
Cross-validation
}
\details{
Generic cross-validation function
}
\examples{
f0 <- function(data,...) lm(...,data)
f1 <- function(data,...) lm(Sepal.Length~Species,data)
f2 <- function(data,...) lm(Sepal.Length~Species+Petal.Length,data)
x <- cv(list(model0=f0,model1=f1,model2=f2),rep=10, data=iris, formula=Sepal.Length~.)
}
\author{
Klaus K. Holst
}
