% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava-package.R
\docType{data}
\name{deprdiag}
\alias{deprdiag}
\title{50 patients from Monash Medical Centre, Melbourne}
\format{
data.frame
}
\source{
Clarke, D. M., Smith, G. C., & Herrman, H. E. (1993). A comparative
  study of screening instruments for mental disorders in general hospital
  patients. International Journal Psychiatry in Medicine, 23, pp. 323-337.

McKenzie et al. (1996). Comparing correlated Kappas by resampling: Is one
level of agreement significantly different from another? J. Psychiat. Res.
30 (6), pp. 483-492.
}
\description{
Diagnosis of depression (DSM-III-R MDD, Dysthymia, Adjustment Disorder with
Depressed Mood, Depression NOS), Beck Depression Inventory (BDI) (Beck et
al., 1961) General Health Questionnaire (GHQ) (Goldberg & Williams, 1988)
}
\keyword{datasets}
