% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devcoords.R
\name{devcoords}
\alias{devcoords}
\title{Returns device-coordinates and plot-region}
\usage{
devcoords()
}
\value{
A \code{list} with elements
\itemize{
\item \code{dev.x1}: device left x-coordinate
\item \code{dev.x2}: device right x-coordinate
\item \code{dev.y1}: device bottom y-coordinate
\item \code{dev.y2}: device top y-coordinate
\item \code{fig.x1}: plot left x-coordinate
\item \code{fig.x2}: plot right x-coordinate
\item \code{fig.y1}: plot bottom y-coordinate
\item \code{fig.y2}: plot top y-coordinate
}
}
\description{
Returns device-coordinates and plot-region
}
\author{
Klaus K. Holst
}
\keyword{hplot}
