% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagtest.R
\name{diagtest}
\alias{diagtest}
\title{Calculate diagnostic tests for 2x2 table}
\usage{
diagtest(table, positive = 2, ...)
}
\arguments{
\item{table}{Table or (matrix/data.frame with two columns)}

\item{positive}{Switch reference}

\item{...}{Additional arguments to lower level functions}
}
\description{
Calculate prevalence, sensitivity, specificity, and positive and
negative predictive values
}
\author{
Klaus Holst
}

