% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsort.R
\name{dsort}
\alias{dsort}
\title{Sort data frame}
\usage{
dsort(data, x, ..., decreasing = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{x}{variable to order by}

\item{...}{additional variables to order by}

\item{decreasing}{sort order (vector of length x)}
}
\value{
data.frame
}
\description{
Sort data according to columns in data frame
}
\examples{
data(data="hubble",package="lava")
dsort(hubble, "sigma")
dsort(hubble, hubble$sigma,"v")
dsort(hubble,~sigma+v)
dsort(hubble,~sigma-v)
}

