% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.formula.R
\name{estimate.formula}
\alias{estimate.formula}
\title{Estimate method for formulas}
\usage{
\method{estimate}{formula}(
  x,
  data,
  weights,
  family = stats::gaussian,
  ...,
  model = "glm",
  lvm = FALSE
)
}
\arguments{
\item{x}{formula specifying the model}

\item{data}{data.frame}

\item{weights}{optional weights}

\item{family}{GLM family (default \code{gaussian})}

\item{...}{Additional arguments to \link{estimate.default}}

\item{model}{character string specifying the model type (default \code{"glm"})}

\item{lvm}{logical, if TRUE use \code{estimate.lvm} instead}
}
\value{
Object of class \code{estimate} (see \link{estimate.default}).
}
\description{
Fit a GLM model specified by a formula and apply \link{estimate.default}
for robust inference.
}
