% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.glm.R
\name{estimate.glm}
\alias{estimate.glm}
\title{Estimate method for GLM objects}
\usage{
\method{estimate}{glm}(x, ...)
}
\arguments{
\item{x}{Fitted \code{glm} object}

\item{...}{Additional arguments to \link{estimate.default}}
}
\value{
Object of class \code{estimate} (see \link{estimate.default}).
}
\description{
Applies \link{estimate.default} to obtain robust inference for a fitted
\code{glm} object.
}
