% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.list.R
\name{estimate.list}
\alias{estimate.list}
\title{Estimate method for lists}
\usage{
\method{estimate}{list}(x, ...)
}
\arguments{
\item{x}{list of model objects or \code{lvm} objects}

\item{...}{Additional arguments to lower level functions}
}
\value{
Object of class \code{estimate.list} (a list of \code{estimate} objects),
or a fitted multi-group model if input is a list of \code{lvm} objects.
}
\description{
Estimate method for lists
}
