% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{fplot
Fast plotting via RGL

Fast plotting via RGL}
\usage{
fplot(x, y, z = NULL, xlab, ylab, ..., z.col = topo.colors(64),
  data = parent.frame(), add = FALSE)
}
\description{
fplot
Fast plotting via RGL

Fast plotting via RGL
}
\examples{
if (interactive()) {
data(iris)
fplot(Sepal.Length ~ Petal.Length+Species, data=iris, size=2, type="s")
}
}

