% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/zgetsas.R
\name{getSAS}
\alias{getSAS}
\title{Read SAS output}
\usage{
getSAS(infile, entry = "Parameter Estimates", ...)
}
\arguments{
\item{infile}{file (csv file generated by ODS)}

\item{entry}{Name of entry to capture}

\item{\dots}{additional arguments to lower level functions}
}
\description{
Run SAS code like in the following:
}
\details{
ODS CSVALL BODY="myest.csv";
proc nlmixed data=aj qpoints=2 dampstep=0.5;
...
run;
ODS CSVALL Close;

and read results into R with:

\code{getsas("myest.csv","Parameter Estimates")}
}
\author{
Klaus K. Holst
}
\seealso{
getMplus
}

