% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{Graph}
\alias{Graph}
\alias{Graph<-}
\title{Extract graph}
\usage{
Graph(x, ...)

Graph(x, ...) <- value
}
\arguments{
\item{x}{Model object}

\item{\dots}{Additional arguments to be passed to the low level functions}

\item{value}{New \code{graphNEL} object}
}
\description{
Extract or replace graph object
}
\examples{

m <- lvm(y~x)
Graph(m)

}
\seealso{
\code{\link{Model}}
}
\author{
Klaus K. Holst
}
\keyword{graphs}
\keyword{models}
