% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lava-package.R
\docType{data}
\name{hubble}
\alias{hubble}
\title{Hubble data}
\format{data.frame}
\source{
Freedman, W. L., et al. 2001, AstroPhysicalJournal, 553, 47.
}
\description{
Velocity (v) and distance (D) measures of 36 Type Ia super-novae from the Hubble
Space Telescope
}
\keyword{datasets}

