% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid}
\alias{iid}
\title{Extract i.i.d. decomposition from model object}
\usage{
iid(x, ...)
}
\arguments{
\item{x}{Model object}

\item{...}{Additional arguments (see the man-page of the IC method)}
}
\value{
Matrix (n x p) with the i.i.d. decomposition of the estimator
(scaled influence function, i.e., the influence function divided by n).
}
\description{
This function extracts
}
