% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-method.R
\name{index.lvm}
\alias{index.lvm}
\alias{index.lvmfit}
\alias{index<-.lvm}
\alias{index<-.lvmfit}
\title{Extract the parameter indicies of a lvm object}
\usage{
\method{index}{lvm}(x, ...)

\method{index}{lvm}(x, ...) <- value
}
\arguments{
\item{x}{object on which to set the index}

\item{...}{further arguments to be passed to or from other methods.}

\item{value}{new index}
}
\description{
Extracts the matrices with indices of model parameters from a
latent variable model (\code{lvm}). Returns a list with
\itemize{
\item A:   Matrix with fixed parameters and ones where parameters are free
\item J:   Manifest variable selection matrix
\item M0:  Index of free regression parameters
\item M1:  Index of free and \emph{unique} regression parameters
\item P:   Matrix with fixed variance parameters and ones where parameters are free
\item P0:  Index of free variance parameters
\item P1:  Index of free and \emph{unique} regression parameters
\item npar.var:   Number of covariance parameters
}
}
\seealso{
\code{\link[=modelPar]{modelPar()}}
}
