% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.sem.R
\name{index}
\alias{index}
\alias{index<-}
\title{Generic method for extract index of an object}
\usage{
index(x, ...)

index(x, ...) <- value
}
\arguments{
\item{x}{object on which to set the index}

\item{...}{further arguments to be passed to or from other methods.}

\item{value}{the new index}
}
\description{
Generic method for extract index of an object

Generic method for setting the index of an object
}
\seealso{
\link{index<-}

\link{index}
}
\keyword{utilities}
