% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{information}
\alias{information}
\title{Extract information matrix}
\usage{
information(x, ...)
}
\arguments{
\item{x}{Model object}

\item{...}{Additional arguments to lower level functions}
}
\value{
Information matrix or its inverse (depending on arguments).
}
\description{
Extract the (observed or expected) information matrix from a model object.
}
