% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse.R
\name{Inverse}
\alias{Inverse}
\title{Generalized matrix inverse}
\usage{
Inverse(
  X,
  tol = lava.options()$itol,
  det = TRUE,
  names = !chol,
  chol = FALSE,
  symmetric = FALSE
)
}
\arguments{
\item{X}{nxn matrix}

\item{tol}{tolerance for pseudo inverse}

\item{det}{logical, if true the determinant is returned}

\item{names}{preserve dimnames}

\item{chol}{use Cholesky decomposition for calculating inverse otherwise SVD}

\item{symmetric}{set to true if matrix is symmetric}
}
\description{
Generalized matrix inverse
}
