% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km}
\alias{km}
\title{Weighted K-means}
\usage{
km(x, mu, data, weights = rep(1, NROW(x)), iter.max = 20, n.start = 5,
  ...)
}
\arguments{
\item{x}{Data (or formula)}

\item{mu}{Initial centers (or number centers chosen randomly among x)}

\item{data}{optional data frmae}

\item{weights}{Optional weights}

\item{iter.max}{Max number of iterations}

\item{n.start}{Number of restarts}

\item{...}{Additional arguments to lower level functions}
}
\description{
Weighted K-means via Lloyd's algorithm
}
\author{
Klaus K. Holst
}
