% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ksmooth}
\alias{ksmooth}
\alias{surface}
\title{Plot/estimate surface}
\usage{
ksmooth(x, data, h, rgl = TRUE, xlab, ylab, zlab = "", gridsize = rep(51L,
  2), ...)
}
\arguments{
\item{x}{formula or data}

\item{data}{data.frame}

\item{h}{bandwidth}

\item{rgl}{If TRUE 'rgl' is used}

\item{xlab}{X label}

\item{ylab}{Y label}

\item{zlab}{Z label}

\item{gridsize}{grid size of kernel smoother}

\item{...}{Additional arguments to graphics routine (persp3d or persp)}
}
\description{
Plot/estimate surface
}
\examples{
f <- function(x,y) 1-sqrt(x^2+y^2)
surface(f,xlim=c(-1,1),alpha=0.9,aspect=c(1,1,0.75))
surface(f,xlim=c(-1,1),clut=heat.colors(128))
##play3d(spin3d(axis=c(0,0,1), rpm=8), duration=5)


surface(function(x) dmvn(x,sigma=diag(2)),c(-3,3),lit=FALSE,smooth=FALSE,box=FALSE,alpha=0.8)
}

