% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava-package.R
\docType{package}
\name{lava-package}
\alias{lava-package}
\alias{lava}
\title{Estimation and simulation of latent variable models}
\description{
Framwork for estimating parameters and simulate data from Latent Variable
Models.
}
\examples{

lava()

}
\author{
Klaus K. Holst Maintainer: <klaus@holst.it>
}
\keyword{package}
