\name{lava-package}
\alias{lava-package}
\alias{lava}
\docType{package}
\title{
  Estimation and simulation of latent variable models
}
\description{
  Framwork for estimating parameters and simulate data from 
  Latent Variable Models.
}
\details{
\tabular{ll}{
Package: \tab lvm\cr
Type: \tab Package\cr
Version: \tab 0.9-5\cr
Date: \tab 2009-05-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Klaus K. Holst
Maintainer: <k.k.holst@biostat.ku.dk>
}
\keyword{ package }
\examples{
lava()
}