% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_calculus.R
\name{merge.estimate}
\alias{merge.estimate}
\title{Merge estimate objects}
\usage{
\method{merge}{estimate}(
  x,
  y,
  ...,
  id,
  paired = FALSE,
  labels = NULL,
  keep = NULL,
  subset = NULL,
  regex = FALSE,
  sep = FALSE,
  drop.ic = FALSE,
  ignore.case = FALSE,
  sort = FALSE
)
}
\arguments{
\item{x}{Object of class \code{estimate}}

\item{y}{Object of class \code{estimate}}

\item{...}{Additional \code{estimate} objects or arguments}

\item{id}{Optional cluster variable}

\item{paired}{If TRUE a paired (matched) analysis is performed}

\item{labels}{Optional character vector of labels for the merged estimates}

\item{keep}{Optional character vector of parameter names to keep}

\item{subset}{Optional character vector of parameter names to subset}

\item{regex}{If TRUE, \code{keep} and \code{subset} are treated as regular expressions}

\item{sep}{Separator used for labeling}

\item{drop.ic}{If TRUE, drop the influence function from the result}

\item{ignore.case}{If TRUE, case is ignored in \code{keep}/\code{subset} matching}

\item{sort}{if true the returned influence function will be sorted according
to the id variables (lexigraphically)}
}
\value{
Object of class \code{estimate} (see \link{estimate.default}).
}
\description{
Merge estimate objects
}
\seealso{
\code{\link[=c.estimate]{c.estimate()}}
}
