% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{meta.analyse.locus}
\alias{meta.analyse.locus}
\title{Re-process locus to meta-analyse of selected phenotypes}
\usage{
meta.analyse.locus(locus, meta.phenos)
}
\arguments{
\item{locus}{Locus object defined using the \code{\link{process.locus}} function.}

\item{meta.phenos}{Phenotypes you want to meta-analyse

#'}
}
\value{
This function returns an object just like that \code{\link{process.locus}} function, containing general locus info, the relevant processed sumstats, and info about the input phenotypes.

\itemize{
    \item id - locus ID
    \item chr/start/stop - locus coordinates
    \item snps - list of locus SNPs
    \item N.snps - number of SNPs
    \item K - number of PCs
    \item delta - PC projected joint SNP effects for each phenotype
    \item sigma - sampling covariance matrix
    \item omega - genetic covariance matrix
    \item omega.cor - genetic correlation matrix
    \item N - vector of average N across locus SNPs for each phenotype
    \item phenos - phenotype IDs
    \item binary - boolean vector indicating whether phentoypes are binary
}
}
\description{
Will combine all elements of the requested phenotypes using standard inverse variance weighting, allowing them to be analysed as a single phenotype via the  multivariate analysis functions.
Note that the univariate test cannot currently be applied to meta-analysed phenotypes, so please do that beforehand on each phenotype individually.
}
