% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napass0.R
\name{na.pass0}
\alias{na.pass0}
\title{Handle Missing Values in Objects}
\usage{
na.pass0(object, na.action = na.omit, ...)
}
\arguments{
\item{object}{data.frame, vector}

\item{na.action}{function used to identify missing values}

\item{...}{additional arguments to lower level functions}
}
\description{
Returns the object with missing data replaced by zeros. This is
  sometimes useful for example when working with inverse probability
  weighting of the complete-case data.
}
\examples{
d <- data.frame(y=c(1,1,NA,2,NA,2), r=c(1,1,0,1,1,1))
na.pass0(d)
glm(y ~ 1, weights=d$r, data=d, na.action=na.pass0)
}
\seealso{
[na.pass], [na.omit], [na.fail]
}
