% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA2x.R
\name{NA2x}
\alias{NA2x}
\alias{x2NA}
\title{Convert to/from NA}
\usage{
NA2x(s, x = 0)
}
\arguments{
\item{s}{The input vector (of arbitrary class)}

\item{x}{The elements to transform into \code{NA} resp. what to transform
\code{NA} into.}
}
\value{
A vector with same dimension and class as \code{s}.
}
\description{
Convert vector to/from NA
}
\examples{
##' 
x2NA(1:10, 1:5)
NA2x(x2NA(c(1:10),5),5)##'
}
\author{
Klaus K. Holst
}
\keyword{manip}
