% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_calculus.R
\name{normcdf}
\alias{normcdf}
\alias{normpdf}
\alias{norminv}
\alias{normcdf.default}
\alias{normpdf.default}
\alias{norminv.default}
\title{Normal distribution}
\usage{
normcdf(q, mean = 0, sd = 1, ...)

\method{normcdf}{default}(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE, ...)

normpdf(x, mean = 0, sd = 1, ...)

\method{normpdf}{default}(x, mean = 0, sd = 1, log = FALSE, ...)

norminv(p, mean = 0, sd = 1, ...)

\method{norminv}{default}(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE, ...)
}
\arguments{
\item{mean}{mean of the normal distribution (numeric or \code{estimate} object)}

\item{sd}{standard deviation of the normal distribution (numeric or
\code{estimate} object)}

\item{...}{additional arguments to lower level functions (e.g., \code{labels})}

\item{lower.tail}{if \code{TRUE} (default) probabilities are
\eqn{P(X\leq x)}, otherwise \eqn{P(X > x)}}

\item{x, q}{vector of quantiles, or an \code{estimate} object}

\item{log, log.p}{if \code{TRUE} probabilities/densities are given as
\eqn{\log(p)}}

\item{p}{vector of probabilities, or an \code{estimate} object}
}
\value{
A numeric vector (default method) or an object of class \code{estimate}.
}
\description{
Density (\code{normpdf}), distribution function (\code{normcdf}), and quantile
function (\code{norminv}) of the normal distribution. In contrast to
\link[stats:dnorm]{stats::dnorm}, \link[stats:pnorm]{stats::pnorm}, and \link[stats:qnorm]{stats::qnorm} these are S3 generic
functions, and methods are available for objects of class \code{estimate} (see
\link{estimate.default}) in which case the delta method is applied to obtain the
influence function (and hence robust standard errors) of the transformed
parameter.
}
\details{
Any of the arguments \code{x}/\code{q}/\code{p}, \code{mean} and \code{sd} may be an \code{estimate}
object. When more than one of them is an \code{estimate} object the objects are
first merged (see \link{merge.estimate}) and the joint delta method is applied,
correctly propagating the dependence between them.
}
\examples{
e <- estimate(coef = c(a = 1, b = 2), vcov = diag(2) * 0.1)
normcdf(e)
normpdf(e)
norminv(normcdf(e)) # identity

## mean and sd may themselves be estimates
mu <- estimate(coef = c(mu = 0.5), vcov = 0.01)
normcdf(e, mean = mu, sd = 2)
}
\seealso{
\link{estimate.default} \link[stats:dnorm]{stats::dnorm}
}
\author{
Klaus Kähler Holst
}
