% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{pcor}
\alias{pcor}
\title{Polychoric correlation}
\usage{
pcor(x, y, X, start, ...)
}
\arguments{
\item{x}{Variable 1}

\item{y}{Variable 2}

\item{X}{Optional covariates}

\item{start}{Optional starting values}

\item{...}{Additional arguments to lower level functions}
}
\description{
Maximum likelhood estimates of polychoric correlations
}
