\name{PD}
\alias{PD}
\title{Dose response...}
\usage{
  PD(model, intercept = 1, slope = 2, prob = NULL, x,
    level = 0.5, ci.level = 0.95, EB = NULL)
}
\arguments{
  \item{model}{}

  \item{intercept}{}

  \item{slope}{}

  \item{prob}{}

  \item{x}{}

  \item{level}{}

  \item{ci.level}{}

  \item{EB}{}
}
\description{
  Dose response calculation
}
\author{
  Klaus K. Holst
}

