% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.estimate.R
\name{plot.estimate}
\alias{plot.estimate}
\title{Plot method for 'estimate' objects}
\usage{
\method{plot}{estimate}(x, f, idx, intercept = FALSE, data, confint = TRUE,
  type = "l", xlab = "x", ylab = "f(x)", col = 1, add = FALSE, ...)
}
\arguments{
\item{x}{estimate object}

\item{f}{function of parameter coefficients and data parsed on to 'estimate'.
If omitted a forest-plot will be produced.}

\item{idx}{Index of parameters (default all)}

\item{intercept}{include intercept in forest-plot}

\item{data}{data.frame}

\item{confint}{Add confidence limits}

\item{type}{plot type ('l')}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{col}{color}

\item{add}{add plot to current device}

\item{...}{additional arguments to lower-level functions}
}
\description{
Plot method for 'estimate' objects
}
