% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.default.R
\name{plot.sim}
\alias{plot.sim}
\alias{density.sim}
\title{Plot sim object}
\usage{
\method{plot}{sim}(x, estimate, se = NULL, true = NULL, names = NULL,
  auto.layout = TRUE, byrow = FALSE, type = "p",
  ask = grDevices::dev.interactive(), line.col = 1, line.lwd = 1.8,
  col = c("gray60", "orange", "darkblue", "seagreen", "darkred"), pch = 16,
  cex = 0.5, lty = 1, true.lty = 2, true.col = "gray70",
  true.lwd = 1.2, legend, legendpos = "topleft", cex.legend = 0.8,
  plot.type = c("multiple", "single"), polygon = TRUE,
  polygon.density = 0, polygon.angle = -45, cex.axis = 0.8, alpha = 0.5,
  rug = TRUE, rug.alpha = 0.5, main, cex.main = 1, equal = FALSE,
  delta = 1.15, ylim = NULL, ylab = "Estimate",
  density.ylab = "Density", density.ylim = NULL, density.xlim = NULL,
  density.plot = TRUE, scatter.plot = TRUE, running.mean = scatter.plot,
  density.alpha = 0.2, border = density.col, density.lty,
  density.col = col, density.lwd = 0.4, xlab = "", ...)
}
\arguments{
\item{x}{sim object}

\item{estimate}{columns with estimates}

\item{se}{columns with standard error estimates}

\item{true}{(optional) vector of true parameter values}

\item{names}{(optional) names of estimates}

\item{auto.layout}{Auto layout (default TRUE)}

\item{byrow}{Add new plots to layout by row}

\item{type}{plot type}

\item{ask}{if TRUE user is asked for input, before a new figure is drawn}

\item{line.col}{line colour}

\item{line.lwd}{line width}

\item{col}{colour (for each estimate)}

\item{pch}{symbol}

\item{cex}{size}

\item{lty}{line type}

\item{true.lty}{true parameter estimate line type}

\item{true.col}{true parameter colour}

\item{true.lwd}{true parameter line width}

\item{legend}{legend}

\item{legendpos}{legend position}

\item{cex.legend}{size of legend text}

\item{plot.type}{'single' or 'multiple' (default)}

\item{polygon}{if TRUE fill the density estimates with colour}

\item{polygon.density}{if non-zero add shading lines}

\item{polygon.angle}{shading lines angle}

\item{cex.axis}{Font size on axis}

\item{alpha}{Semi-transparent level (1: non-transparent, 0: full)}

\item{rug}{if TRUE add rug representation of data to x-axis}

\item{rug.alpha}{rug semi-transparency level}

\item{main}{Main title}

\item{cex.main}{Size of title font}

\item{equal}{}

\item{delta}{}

\item{ylim}{y-axis limits}

\item{ylab}{}

\item{density.ylab}{y label for density plots}

\item{density.ylim}{y-axis limits for density plots}

\item{density.xlim}{x-axis limits for density plots}

\item{density.plot}{if TRUE add density plot}

\item{scatter.plot}{if TRUE add scatter plot}

\item{running.mean}{if TRUE add running average estimate to scatter plot}

\item{density.alpha}{semi-transparency of}

\item{border}{}

\item{density.lty}{}

\item{density.col}{}

\item{density.lwd}{}

\item{xlab}{}

\item{...}{}
}
\description{
Plot sim object
}
\examples{
n <- 1000
val <- cbind(est1=rnorm(n,sd=1),est2=rnorm(n,sd=0.2),est3=rnorm(n,1,sd=0.5),
             sd1=runif(n,0.8,1.2),sd2=runif(n,0.1,0.3),sd3=runif(n,0.25,0.75))

plot.sim(val,estimate=c(1,2),true=c(0,0),se=c(4,5),equal=TRUE)
plot.sim(val,estimate=c(1,3),true=c(0,1),se=c(4,6),density.xlim=c(-3,3),ylim=c(-3,3))
plot.sim(val,estimate=c(1,2),true=c(0,0),se=c(4,5),equal=TRUE,plot.type="single")
plot.sim(val,estimate=c(1),se=c(4,5,6),plot.type="single")
plot.sim(val,estimate=c(1,2,3),equal=TRUE)
plot.sim(val,estimate=c(1,2,3),equal=TRUE,byrow=TRUE)
plot.sim(val,estimate=c(1,2,3),plot.type="single")
plot.sim(val,estimate=1,se=c(3,4,5),plot.type="single")

density.sim(val,estimate=c(1,2,3),polygon.density=c(0,10,10),polygon.angle=c(0,45,-45))
}
