% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.default.R
\name{Print}
\alias{Print}
\title{Generic print method}
\usage{
Print(x, n = 5, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{object to print}

\item{n}{number of rows to show from top and bottom of tabular data}

\item{digits}{precision}

\item{...}{additional arguments to print method}
}
\description{
Nicer print method for tabular data. Falls back to standard print method for
all other data types.
}
