% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_eqtl.r
\name{process.eqtl.locus}
\alias{process.eqtl.locus}
\title{Process eQTL locus info}
\usage{
process.eqtl.locus(
  gene,
  eqtl.input,
  phenos = NULL,
  min.K = 2,
  prune.thresh = 99,
  max.prop.K = 0.75,
  drop.failed = T
)
}
\arguments{
\item{gene}{Gene to analyse, from the available genes in eqtl.input$current.genes. Can either be a gene name, or a numeric index.}

\item{eqtl.input}{Input object created with the \code{\link{process.input}} function and modified with \code{\link{process.eqtl.input}}, containing
relevant summary statistics and related info (e.g. sample overlap, case/control ratio)}

\item{phenos}{Subset of phenotypes from the input object to process. If NULL (default), all phenotypes will be processed}

\item{min.K}{Minimum number of PCs required to process locus (cannot be less than two). If this criterion is not met, the function will fail and the locus cannot be analysed.}

\item{prune.thresh}{PC pruning threshold governing the maximum number of PCs to retain.}

\item{max.prop.K}{Upper bound on retained number of PCs as proportion of lowest sample size of input data.}

\item{drop.failed}{Determines if failed phenotypes are removed from the output object (default) or retained.
PCs are selected as such that the cumulative proportion of variance explained is at least that of the threshold (set to 99 percent by default).}
}
\value{
Returns an environment containing general locus info, the processed locus related sumstats, and parameters required for analysis. If the function fails (e.g. due to too few SNPs), it will return NULL.
If processing fails for specific phenotypes, only the successful phenotypes will be returned (unless the drop.failed argument is set to true).

\itemize{
    \item id - locus ID
    \item chr / start / stop - locus coordinates
    \item snps - list of locus SNPs
    \item n.snps - number of SNPs within locus
    \item K - number of PCs retained
    \item delta - PC projected joint SNP effects
    \item sigma - sampling covariance matrix
    \item omega - genetic covariance matrix
    \item omega.cor - genetic correlation matrix
    \item N - vector of average N across locus SNPs for each phenotype
    \item phenos - phenotype IDs
    \item binary - boolean vector indicating whether phenotypes are binary
    \item h2.obs - observed local heritability
    \item h2.latent - estimated local population heritability (only relevant for binary phenotypes; requires population prevalence to be specified in the input info file)
    \item failed - boolean vector indicating whether phenotypes failed during processing (only present if drop.failed=F)
}
}
\description{
Same as \code{\link{process.locus}}, but for use with eQTL data. Processes the summary statistics for the eQTL data and
all phenotypes within the specified genes and computes all parameters necessary for analysis.
}
