% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_eqtl.r
\name{process.eqtl.text.input}
\alias{process.eqtl.text.input}
\title{Process eQTL input from text input files}
\usage{
process.eqtl.text.input(
  gwas.input,
  eqtl.file.spec,
  chromosomes = "all",
  sample.size = NULL
)
}
\arguments{
\item{gwas.input}{GWAS input object generated using \code{\link{process.input}} (object is modified by function).}

\item{eqtl.file.spec}{File name specifier for eQTL summary statistics input files, with a placeholder [CHR] to indicate the chromosome number in the file name; eg. "my_folder/eqtl.chr[CHR].stats"}

\item{chromosomes}{Specifies which chromosomes to load the data for (loading all chromosomes may be time-consuming).
Can be set to 'all' to load all chromosomes, 'auto' to load all autosomal chromosomes, or a vector of chromosome numbers or single chromosome number (the X chromosome can be specified as 'X' or 23).}

\item{sample.size}{If set, used as sample size value for each SNP in the input files instead of N column.}
}
\value{
No return value, the gwas.input argument is modified in place. An 'eqtl' phenotype is added into the existing gwas.input structure, and the following
fields are added to gwas.input.
\itemize{
    \item current.chromosomes - list of currently loaded chromosomes
    \item current.genes - vector of available genes
    \item current.stats - summary statistics for all loaded genes
}
}
\description{
Modifies the GWAS input object, appending eQTL (or sQTL) summary statistics and gene annotation to an existing GWAS input from plain-text input files. Input files should follow the same format
as summary statistics input files for other phenotypes, with an additional column GENE to indicate which gene the association results on each line belong to. Input files must be provided separately
per chromosome.
}
