% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{read.loci}
\alias{read.loci}
\title{Read in locus file}
\usage{
read.loci(loc.file)
}
\arguments{
\item{loc.file}{Name of locus file}
}
\description{
This function reads in the locus file, defining the locus boundaries using either coordinates (headers: 'CHR', 'START', 'STOP') and/or ';' separated SNP lists (header: 'SNPS').
A locus ID column is also required (header: 'LOC').\cr
\cr
If both coordinates and SNP list columns are provided, only the SNP lists will be used for subsetting the reference data (this can be convenient if the SNP coordinates are based on a different GRChX version than the reference).\cr
\cr
If a SNP column is provided, please remove any columns with 0 SNPs in advance
}
