% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdiag.R
\name{revdiag}
\alias{offdiag}
\alias{offdiag<-}
\alias{revdiag}
\alias{revdiag<-}
\title{Create/extract 'reverse'-diagonal matrix or off-diagonal elements}
\usage{
revdiag(x,...)
offdiag(x,...)

revdiag(x,...) <- value
offdiag(x,...) <- value
}
\arguments{
\item{x}{vector}

\item{value}{For the assignment function the values to put in the diagonal}

\item{\dots}{additional arguments to lower level functions}
}
\description{
Create/extract 'reverse'-diagonal matrix or off-diagonal elements
}
\author{
Klaus K. Holst
}

