% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotate2}
\alias{rotate2}
\alias{rot2D,}
\alias{rot3D}
\title{Performs a rotation in the plane}
\usage{
rotate2(x, theta = pi)
}
\arguments{
\item{x}{Matrix to be rotated (2 times n)}

\item{theta}{Rotation in radians}
}
\value{
Returns a matrix of the same dimension as \code{x}
}
\description{
Performs a rotation in the plane
}
\examples{
rotate2(cbind(c(1,2),c(2,1)))
}
\author{
Klaus K. Holst
}
\keyword{hplot}
