% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run.bivar}
\alias{run.bivar}
\title{Bivariate local genetic correlation analysis}
\usage{
run.bivar(
  locus,
  phenos = NULL,
  adap.thresh = c(1e-04, 1e-06),
  p.values = T,
  CIs = T,
  param.lim = 1.25
)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function, containing all the relevant parameters for the phenotypes of interest}

\item{phenos}{Optional parameter specifying subset of phenotypes. If NULL, all phenotypes in the locus object will be analysed}

\item{adap.thresh}{Adaptive iteration procedure for permutation p-values that increases the number of permutations for lower p-values. Can be set turned of by setting to NULL.}

\item{p.values}{Set to F if p-values are not desired.}

\item{CIs}{Set to F to if confidence intervals are not desired.}

\item{param.lim}{The +- threshold at which estimated parameters are considered to be too far out of bounds (and will be set to NA)}
}
\description{
Performs bivariate local genetic correlation analysis between two phenotypes for a single locus.
If more than one phenotype is specified, the last phenotype will be treated as the phenotype of interest, 
and separate bivariate tests will be performed between this phenotype and all others.
Note: this test is symmetric, which phenotype is considered predictor/outcome doesn't matter.
}
