% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run.partial.cor}
\alias{run.partial.cor}
\title{Local partial genetic correlation analysis}
\usage{
run.partial.cor(
  locus,
  phenos = NULL,
  adap.thresh = c(1e-04, 1e-06),
  p.values = T,
  CIs = T,
  max.r2 = 0.95,
  param.lim = 1.25
)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function. Contains all the relevant parameters and processed sum-stats for the phenotypes of interest}

\item{phenos}{Optional argument specifying subset of phenotypes to analyse, and/or their order. If NULL, all phenotypes in the locus object will be analysed (in the order of the locus object). NOTE: the first two are p1 & p2 (the phenotypes of interest), the remaining are Z}

\item{p.values}{Set to F to suppress p-values}

\item{CIs}{Set to F to suppress 95\% confidence intervals}

\item{max.r2}{Max r2 threshold for the regression of phen1 ~ Z and phen2 ~ Z. If any of these r2's are too high, the partial correlation becomes unstable, and analysis is therefore aborted.}

\item{param.lim}{The +- threshold at which estimated parameters are considered to be too far out of bounds. If the estimated parameter exceeds this threshold, it is considered unreliable and will be set to NA. 

#' @return Data frame with the columns:
\itemize{
    \item phen1 / phen2 - analysed phenotypes
    \item rho - standardised coefficient for the partial genetic correlation
    \item rho.lower / rho.upper - 95\% confidence intervals for rho
    \item r2 - proportion of partial variance in genetic signal for phen1 explained by phen2 (and vice versa)
    \item r2.lower / r2.upper - 95\% confidence intervals for the r2
    \item p - permutation p-values for the partial genetic correlation
}}

\item{adapt.thresh}{The thresholds at which to increase the number of permutations for the p-value generation. 
Default number of permutations is 1e+4, but will be increased to 1e+5, and 1e+6 as p-values fall below the respective thresholds.
If set to NULL, the maximum number of permutations is capped at the default (Note: this significantly speeds up the analysis, but results in poor accuracy for low p-values)}
}
\description{
Will perform a local partial genetic correlation between the first two phenotypes (phen1, phen2) conditioned on the rest (Z). 
Phenotype order is based on that within the locus object by default, but can be changed by passing a phenotype vector with the desired order / phenotype subset to the 'phenos' argument.
}
