% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run.univ.bivar}
\alias{run.univ.bivar}
\title{Perform both univariate and bivariate tests, with filtering based on univariate signal}
\usage{
run.univ.bivar(
  locus,
  phenos = NULL,
  target = NULL,
  univ.thresh = 0.05,
  adap.thresh = c(1e-04, 1e-06),
  p.values = T,
  CIs = T,
  param.lim = 1.25,
  cap.estimates = T
)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function. Contains all the relevant parameters and processed sum-stats for the phenotypes of interest}

\item{phenos}{Subset of phenotypes to analyse, and/or their order. If NULL, all phenotypes in the locus object will be analysed (in the order listed within the locus object)}

\item{target}{Target phenotype of interest. If NULL, bivariate correlations between all pairs of phenotypes will be computed; 
Otherwise, this will only be done between the target phenotype and all other phenotypes.}

\item{univ.thresh}{P-value threshold for the univariate test, used to determine whether phenotypes exhibit sufficient local heritability for a bivariate test}

\item{adap.thresh}{The thresholds at which to increase the number of iterations for the p-value generation. 
Default number of iterations is 1e+4, but will be increased to 1e+5, and 1e+6 as p-values fall below the respective thresholds.
If set to NULL, the maximum number of iterations is capped at the default (Note: this significantly speeds up the analysis, but results in poor accuracy for low p-values)}

\item{p.values}{Set to F to suppress p-values}

\item{CIs}{Set to F to suppress 95\% confidence intervals}

\item{param.lim}{The +- threshold at which estimated parameters are considered to be too far out of bounds. If the estimated parameter exceeds this threshold, it is considered unreliable and will be set to NA.}
}
\value{
List containing the results from the univariate and bivariate tests (see ?run.univ() and ?run.bivar() for more info)
}
\description{
Will only perform bivariate test for all phenotypes that pass the univariate significance threshold. By default, the bivariate test will be performed for all combinations of phenotypes in the locus, 
but this can be modified using the 'phenos' and 'target' arguments (see below)
}
