% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run.univ}
\alias{run.univ}
\title{Run univariate analysis}
\usage{
run.univ(locus, phenos = NULL, var = F)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function. Contains all the relevant parameters and processed sum-stats for the phenotypes of interest}

\item{phenos}{Subset of phenotypes to analyse, and/or their order. If NULL, all phenotypes in the locus object will be analysed (in the order listed within the locus object)}

\item{var}{Set to T to return variance estimate}
}
\value{
Data frame with the columns:
\itemize{
    \item phen - analysed phenotypes
    \item var - local genetic variance
    \item h2.obs - observed local heritability
    \item h2.latent - estimated local population heritability (only relevant for binary phenotypes; requires population prevalence to be specified in input info file)
    \item p - p-values from the univariate test (F-test for continuous, Chi-sq for binary)
}
}
\description{
Tests the univariate local genetic signal (i.e. the local heritability) for all phenotypes.
}
