% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Extract score function}
\usage{
score(x, ...)
}
\arguments{
\item{x}{Model object}

\item{...}{Additional arguments to lower level functions}
}
\value{
Matrix of score/gradient values. Rows correspond to
observations (if \code{indiv=TRUE}) or a single row, columns to parameters.
}
\description{
Extract the score function (gradient of the log-likelihood) from a
model object.
}
