% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_eqtl.r
\name{set.chromosomes}
\alias{set.chromosomes}
\title{Load gene annotation for eQTL input object}
\usage{
set.chromosomes(gwas.input, chromosomes)
}
\arguments{
\item{gwas.input}{GWAS input object loaded using \code{\link{process.eqtl.input.preprocessed}} (object is modified by function).}

\item{chromosomes}{Specifies which chromosomes to load the annotation data for (loading all chromosomes may be time-consuming).
Can be set to 'all' to load all chromosomes, 'auto' to load all autosomal chromosomes, or a vector of chromosome numbers or single chromosome number (the X chromosome can be specified as 'X' or 23).}
}
\value{
No return value, the gwas.input argument is modified in place. The following fields are changed
\itemize{
    \item eqtl.genes - list containing the names and mapped SNPs for all genes in the eQTL data set
    \item current.chromosomes - list of currently loaded chromosomes
    \item current.genes - vector of genes available for the currently loaded tissue
}
}
\description{
Load the eQTL gene annotation for the specified chromosomes into the gwas.input object. Previously loaded gene annotation is replaced
}
