% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_eqtl.r
\name{set.tissue}
\alias{set.tissue}
\title{Initialize specific tissue for eQTL input object}
\usage{
set.tissue(gwas.input, tissue)
}
\arguments{
\item{gwas.input}{GWAS input object preprocessed using \code{\link{process.eqtl.input}} (object is modified by function).}

\item{tissue}{Tissue to load summary statistics for.}
}
\value{
No return value, the gwas.input argument is modified in place. The following fields are changed
\itemize{
    \item current.tissue - name of the currently loaded tissue
    \item current.genes - vector of genes available for the currently loaded tissue
    \item current.stats - summary statistics for the currently loaded tissue
}
}
\description{
Modifies the eQTL input object prepared by the process.eqtl.input() function, loading and initializing summary statistics for the specified tissue. Removes data loaded
for previously loaded tissue (if any).
}
