% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spaghetti.R
\name{spaghetti}
\alias{spaghetti}
\title{Spaghetti plot}
\usage{
spaghetti(formula, data, id, type = "l", lty = 1, col = Col(1),
  trend = FALSE, trend.col = "darkblue", trend.alpha = 0.2,
  trend.lwd = 3, xlab = "Time", ylab = "", ...)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{id}{}

\item{type}{}

\item{lty}{}

\item{col}{}

\item{trend}{}

\item{trend.col}{}

\item{trend.alpha}{}

\item{trend.lwd}{}

\item{xlab}{}

\item{ylab}{}

\item{...}{}
}
\description{
Spaghetti plot for longitudinal data
}
\examples{
\dontrun{
library(mets)
K <- 5
y <- "y"\%++\%seq(K)
m <- lvm()
regression(m,y=y,x=~u) <- 1
regression(m,y=y,x=~s) <- seq(K)-1
regression(m,y=y,x=~x) <- "b"
d <- sim(m,500)
dd <- fast.reshape(d);
dd$num <- dd$num+rnorm(nrow(dd),sd=0.5) ## Unbalance
spaghetti(y~num,dd,id="id",lty=1,col=Col(1,.4),trend=TRUE,trend.col="darkblue")
}
}
\author{
Klaus K. Holst
}

