% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.default.R
\name{summary.sim}
\alias{summary.sim}
\title{Summary method for 'sim' objects}
\usage{
\method{summary}{sim}(
  object,
  estimate = NULL,
  se = NULL,
  confint = !is.null(se) && !is.null(true),
  true = NULL,
  fun,
  names = NULL,
  unique.names = TRUE,
  minimal = FALSE,
  level = 0.95,
  quantiles = c(0, 0.025, 0.5, 0.975, 1),
  ...
)
}
\arguments{
\item{object}{sim object}

\item{estimate}{(optional) columns with estimates}

\item{se}{(optional) columns with standard error estimates}

\item{confint}{(optional) list of pairs of columns with confidence limits}

\item{true}{(optional) vector of true parameter values}

\item{fun}{(optional) summary function}

\item{names}{(optional) names of}

\item{unique.names}{if TRUE, unique.names will be applied to column names}

\item{minimal}{if TRUE, minimal summary will be returned}

\item{level}{confidence level}

\item{quantiles}{quantiles}

\item{...}{additional levels to lower-level functions}
}
\description{
Summary method for 'sim' objects
}
