% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Trace operator}
\usage{
tr(x, ...)
}
\arguments{
\item{x}{Square numeric matrix}

\item{\dots}{Additional arguments to lower level functions}
}
\value{
\code{numeric}
}
\description{
Calculates the trace of a square matrix.
}
\examples{

tr(diag(1:5))
}
\author{
Klaus K. Holst
}
\seealso{
\code{\link{crossprod}}, \code{\link{tcrossprod}}
}
\keyword{algebra}
\keyword{math}

