% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim string of (leading/trailing/all) white spaces}
\usage{
trim(x, all = FALSE, ...)
}
\arguments{
\item{x}{String}

\item{all}{Trim all whitespaces?}

\item{\dots}{additional arguments to lower level functions}
}
\description{
Trim string of (leading/trailing/all) white spaces
}
\author{
Klaus K. Holst
}
