% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_ic.R
\name{var_ic}
\alias{var_ic}
\title{Variance based on influence function}
\usage{
var_ic(x, ...)
}
\arguments{
\item{x}{Influence function matrix (observations x parameters), typically
obtained from \link{IC}.}

\item{...}{Additional arguments (currently not used)}
}
\value{
Covariance matrix (p x p).
}
\description{
Computes the empirical variance of the influence function, providing a
sandwich-type variance estimator.
}
