% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vec.R
\name{vec}
\alias{vec}
\title{vec operator}
\usage{
vec(x, matrix = FALSE, sep = ".", ...)
}
\arguments{
\item{x}{Array}

\item{matrix}{If TRUE a row vector (matrix) is returned}

\item{sep}{Seperator}

\item{...}{Additional arguments}
}
\description{
vec operator
}
\details{
Convert array into vector
}
\author{
Klaus Holst
}

