% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkm.R
\name{wkm}
\alias{wkm}
\title{Weighted K-means}
\usage{
wkm(
  x,
  mu,
  data,
  weights = rep(1, NROW(x)),
  iter.max = 20,
  n.start = 5,
  init = "kmpp",
  ...
)
}
\arguments{
\item{x}{Data (or formula)}

\item{mu}{Initial centers (or number centers chosen randomly among x)}

\item{data}{optional data frmae}

\item{weights}{Optional weights}

\item{iter.max}{Max number of iterations}

\item{n.start}{Number of restarts}

\item{init}{method to create initial centres (default kmeans++)}

\item{...}{Additional arguments to lower level functions}
}
\description{
Weighted K-means via Lloyd's algorithm
}
\examples{
## Two well-separated Gaussian blobs in 2-D
set.seed(1)
x <- rbind(matrix(rnorm(100, mean = -3), ncol = 2),
           matrix(rnorm(100, mean =  3), ncol = 2))
res <- wkm(x, mu = 2)
table(res$cluster)
res$center

## Supply explicit initial centers (as a list)
res2 <- wkm(x, mu = list(c(-3, -3), c(3, 3)))

## Weighted clustering: up-weight the second blob
w <- c(rep(1, 50), rep(10, 50))
res3 <- wkm(x, mu = 2, weights = w)

## Formula interface on a data.frame
wkm(~ Sepal.Length + Sepal.Width, data = iris, mu = 3)
}
\author{
Klaus K. Holst
}
