% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapvec.R
\name{wrapvec}
\alias{wrapvec}
\title{Wrap vector}
\usage{
wrapvec(x, delta = 0L, ...)
}
\arguments{
\item{x}{Vector or integer}

\item{delta}{Shift}

\item{...}{Additional parameters}
}
\description{
Wrap vector
}
\examples{
wrapvec(5,2)
}

