% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.rrw.R
\name{binomial.rd}
\alias{binomial.rd}
\alias{binomial.rr}
\title{Define constant risk difference or relative risk association for binary exposure}
\usage{
binomial.rd(
  x,
  response,
  exposure,
  target.model,
  nuisance.model,
  exposure.model = binomial.lvm(),
  ...
)
}
\arguments{
\item{x}{model}

\item{response}{response variable (character or formula)}

\item{exposure}{exposure variable (character or formula)}

\item{target.model}{variable defining the linear predictor for the target model}

\item{nuisance.model}{variable defining the linear predictor for the nuisance model}

\item{exposure.model}{model for exposure (default binomial logit link)}

\item{...}{additional arguments to lower level functions}
}
\description{
Set up model as defined in Richardson, Robins and Wang (2017).
}
