% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleinput.R
\name{binomial.rd}
\alias{binomial.rd}
\title{Binomial risk difference regression}
\usage{
binomial.rd(x, response, exposure, target.model, nuisance.model,
  exposure.model = binomial.lvm(), ...)
}
\arguments{
\item{x}{model}

\item{response}{response variable}

\item{exposure}{exposure variable}

\item{target.model}{design matrix for target model (relative difference)}

\item{nuisance.model}{design matrix for nuisance model (odds product)}

\item{exposure.model}{design matrix for propensity model}

\item{...}{additional arguments to lower level functions}
}
\description{
Implementation of the model by Richardson, Robins and Wang (2017) for risk difference
}
