% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Generic bootstrap method}
\usage{
bootstrap(x, ...)
}
\arguments{
\item{x}{Model object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for calculating bootstrap statistics
}
\author{
Klaus K. Holst
}
\seealso{
\code{bootstrap.lvm} \code{bootstrap.lvmfit}
}

