% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootstrap}
\alias{bootstrap}
\title{Generic bootstrap method}
\usage{
bootstrap(x, ...)
}
\arguments{
\item{x}{Model object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for calculating bootstrap statistics
}
\author{
Klaus K. Holst
}
\seealso{
\code{bootstrap.lvm} \code{bootstrap.lvmfit}
}

