% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{calcium}
\alias{calcium}
\title{Longitudinal Bone Mineral Density Data}
\format{A data.frame containing 560 (incomplete) observations. The 'person'
column defines the individual girls of the study with measurements at
visiting times 'visit', and age in years 'age' at the time of visit. The
bone mineral density variable is 'bmd' (g/cm^2).}
\source{
Vonesh & Chinchilli (1997), Table 5.4.1 on page 228.
}
\description{
Bone Mineral Density Data consisting of 112 girls randomized to receive
calcium og placebo. Longitudinal measurements of bone mineral density
(g/cm^2) measured approximately every 6th month in 3 years.
}
\keyword{datasets}

