% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{children}
\alias{children}
\alias{parents}
\alias{ancestors}
\alias{descendants}
\alias{roots}
\alias{sinks}
\alias{adjMat}
\alias{edgeList}
\title{Extract children or parent elements of object}
\usage{
children(object, ...)
}
\arguments{
\item{object}{Object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for memberships from object (e.g. a graph)
}
\author{
Klaus K. Holst
}
