% Generated by roxygen2 (4.0.0): do not edit by hand
\name{correlation}
\alias{correlation}
\title{Generic method for extracting correlation coefficients of model object}
\usage{
correlation(x, ...)
}
\arguments{
\item{x}{Object}

\item{\dots}{Additional arguments}
}
\description{
Generic correlation method
}
\author{
Klaus K. Holst
}

