% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/zdevcoords.R
\name{devcoords}
\alias{devcoords}
\title{Returns device-coordinates and plot-region}
\usage{
devcoords()
}
\value{
A \code{list} with elements
 \item{dev.x1}{Device: Left x-coordinate}
 \item{dev.x2}{Device: Right x-coordinate}
 \item{dev.y1}{Device Bottom y-coordinate}
 \item{dev.y2}{Device Top y-coordinate}
 \item{fig.x1}{Plot: Left x-coordinate}
 \item{fig.x2}{Plot: Right x-coordinate}
 \item{fig.y1}{Plot: Bottom y-coordinate}
 \item{fig.y2}{Plot: Top y-coordinate}
}
\description{
Returns device-coordinates and plot-region
}
\author{
Klaus K. Holst
}
\keyword{hplot}

