% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dsort.R
\name{dsort}
\alias{dsort}
\title{Sort data frame}
\usage{
dsort(data, x, ...)
}
\arguments{
\item{data}{Data frame}

\item{x}{variable to order by}

\item{...}{additional variables to order by}
}
\value{
data.frame
}
\description{
Sort data according to columns in data frame
}
\examples{
data(hubble)
dsort(hubble, "sigma")
dsort(hubble, hubble$sigma,"v")
dsort(hubble,~sigma+v)
}

