% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{dsort}
\alias{dsort}
\title{Sort data frame}
\usage{
dsort(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{Arguments to order}
}
\value{
data.frame
}
\description{
Sort data according to columns in data frame
}
\examples{
data(hubble)
dsort(hubble, "sigma")
dsort(hubble, hubble$sigma)
dsort(hubble,sigma,v)
}

