% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.default.R
\name{estimate.array}
\alias{estimate.array}
\alias{estimate.data.frame}
\title{Estimate parameters and influence function.}
\usage{
\method{estimate}{array}(x, type = "mean", probs = 0.5, ...)
}
\arguments{
\item{x}{numeric matrix}

\item{type}{target parameter ("mean", "variance", "quantile")}

\item{probs}{numeric vector of probabilities (for type="quantile")}

\item{...}{Additional arguments to lower level functions (i.e., stats::density.default when type="quantile")}
}
\description{
Estimate parameters for the sample mean, variance, and quantiles
}
