% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Expand}
\alias{Expand}
\title{Create a Data Frame from All Combinations of Factors}
\usage{
Expand(x, ...)
}
\arguments{
\item{x}{Data.frame}

\item{...}{vectors, factors or a list containing these}
}
\description{
Create a Data Frame from All Combinations of Factors
}
\details{
Simple wrapper of the 'expand.grid' function.  If x is a table
then a data frame is returned with one row pr individual
observation.
}
\examples{
dd <- Expand(iris, Sepal.Length=2:8, Species=c("virginica","setosa"))
summary(dd)

T <- with(warpbreaks, table(wool, tension))
Expand(T)
}
\author{
Klaus K. Holst
}

