% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{fplot
Faster plot via RGL

Faster plot via RGL}
\usage{
fplot(x, y, z = NULL, xlab, ylab, ..., z.col = topo.colors(64),
  data = parent.frame(), add = FALSE)
}
\arguments{
\item{x}{X variable}

\item{y}{Y variable}

\item{z}{Z variable (optional)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arggument to lower level plot functions}

\item{z.col}{Color}

\item{data}{data.frame}

\item{add}{If TRUE use current active device}
}
\description{
fplot
Faster plot via RGL

Faster plot via RGL
}
\examples{
if (interactive()) {
data(iris)
fplot(Sepal.Length ~ Petal.Length+Species, data=iris, size=2, type="s")
}
}

