% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{fplot}
\usage{
fplot(
  x,
  y,
  z = NULL,
  xlab,
  ylab,
  ...,
  z.col = topo.colors(64),
  data = parent.frame(),
  add = FALSE,
  aspect = c(1, 1),
  zoom = 0.8
)
}
\arguments{
\item{x}{X variable}

\item{y}{Y variable}

\item{z}{Z variable (optional)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arggument to lower-level plot functions}

\item{z.col}{color (use argument alpha to set transparency)}

\item{data}{data.frame}

\item{add}{if TRUE use current active device}

\item{aspect}{aspect ratio}

\item{zoom}{zoom level}
}
\description{
Faster plot via RGL
}
\examples{
if (interactive()) {
data(iris)
fplot(Sepal.Length ~ Petal.Length+Species, data=iris, size=2, type="s")
}
}
