% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getMplus}
\alias{getMplus}
\title{Read Mplus output}
\usage{
getMplus(infile = "template.out", coef = TRUE, ...)
}
\arguments{
\item{infile}{Mplus output file}

\item{coef}{Coefficients only}

\item{\dots}{additional arguments to lower level functions}
}
\description{
Read Mplus output files
}
\author{
Klaus K. Holst
}
\seealso{
getSAS
}

