\name{gof}
\alias{gof}
\title{ Calculates various GOF statistics for model object }
\description{
  Calculates various GOF statistics for model object including global
  chi-squared test statistic and AIC.
}
\usage{
gof(object, ...)
}
\arguments{
  \item{object}{ Model object }
  \item{\dots}{}
}
\value{
  A \code{htest}-object.
}
\author{ Klaus K. Holst }

\keyword{ methods }
\keyword{ models }
