% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC.default}
\alias{IC.default}
\alias{IC}
\alias{var_ic}
\title{Extract influence function from model object}
\usage{
\method{IC}{default}(x, bread, id = NULL, ...)
}
\arguments{
\item{x}{model object}

\item{bread}{(optional) Inverse of derivative of mean score function}

\item{id}{(optional) id/cluster variable}

\item{...}{additional arguments}
}
\description{
Extract i.i.d. decomposition (influence function) from model object
}
\examples{
m <- lvm(y~x+z)
distribution(m, ~y+z) <- binomial.lvm("logit")
d <- sim(m,1e3)
g <- glm(y~x+z,data=d,family=binomial)
var_ic(IC(g))
}
