% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava-package.R
\docType{package}
\name{lava-package}
\alias{lava-package}
\alias{lava}
\title{Estimation and simulation of latent variable models}
\description{
A general implementation of Structural Equation Models wth latent variables
(MLE, 2SLS, and composite likelihood estimators) with both continuous,
censored, and ordinal outcomes (Holst and Budtz-Joergensen (2013)
<doi:10.1007/s00180-012-0344-y>). Mixture latent variable models and
non-linear latent variable models (Holst and Budtz-Joergensen (2020)
<doi:10.1093/biostatistics/kxy082>). The package also provides methods for
graph exploration (d-separation, back-door criterion), simulation of general
non-linear latent variable models, and estimation of influence functions for
a broad range of statistical models.##'
}
\examples{

lava()

}
\author{
Klaus K. Holst Maintainer: <klaus@holst.it>
}
\keyword{package}
