% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{lava-package}
\alias{lava}
\alias{lava-package}
\title{Estimation and simulation of latent variable models}
\description{
Framwork for estimating parameters and simulate data from Latent Variable
Models.
}
\examples{
lava()
}
\author{
Klaus K. Holst Maintainer: <k.k.holst@biostat.ku.dk>
}
\keyword{package}

