% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Model}
\alias{Model}
\alias{Model<-}
\title{Extract model}
\usage{
Model(x, ...)

Model(x, ...) <- value
}
\arguments{
\item{x}{Fitted model}

\item{value}{New model object (e.g. \code{lvm} or \code{multigroup})}

\item{\dots}{Additional arguments to be passed to the low level functions}
}
\value{
Returns a model object (e.g. \code{lvm} or \code{multigroup})
}
\description{
Extract or replace model object
}
\examples{
m <- lvm(y~x)
e <- estimate(m, sim(m,100))
Model(e)
}
\author{
Klaus K. Holst
}
\seealso{
\code{\link{Graph}}
}
\keyword{models}

