% Generated by roxygen2 (4.0.0): do not edit by hand
\name{\%ni\%}
\alias{\%ni\%}
\title{Matching operator (x not in y) oposed to the \code{\%in\%}-operator (x in y)}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{vector}

\item{y}{vector of same type as \code{x}}
}
\value{
A logical vector.
}
\description{
Matching operator
}
\examples{
1:10 \%ni\% c(1,5,10)
}
\author{
Klaus K. Holst
}
\seealso{
\code{\link{match}}
}
\keyword{misc}
\keyword{utilities}

