% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/parpos.R
\name{parpos}
\alias{parpos}
\title{Generic method for finding indeces of model parameters}
\usage{
parpos(x, ...)
}
\arguments{
\item{x}{Model object}

\item{\dots}{Additional arguments}
}
\description{
Generic method for finding indeces of model parameters
}
\author{
Klaus K. Holst
}

