% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{process.input}
\alias{process.input}
\title{Process input}
\usage{
process.input(input.info.file, sample.overlap.file, ref.prefix, phenos = NULL)
}
\arguments{
\item{input.info.file}{Name of info file containing 
phenotype IDs ('phenotype'), N cases ('cases'), N controls ('controls'), sumstats file ('filename').
For continuous phenotypes, the number of controls should be set to 0.}

\item{sample.overlap.file}{Name if file with sample overlap information.
Can be set to NULL if there is no overlap}

\item{ref.prefix}{Prefix of reference genotype data in plink format (*.bim, *.bed, *.fam)}

\item{phenos}{A vector of phenotype IDs can be provided if only a subset of phenotypes are desired
(if NULL, all phenotypes in the input info file will be processed). 
This can be convenient if a subset from a larger number of phenotypes are analysed, as only a single input.info / sample overlap file needs to be created.}
}
\value{
An object containing processed input data and related info
\itemize{
    \item info - the processed input info file. Columns added during processing: 
    \itemize{
       \item N = cases + controls
       \item prop_cases = cases / N
       \item binary = prop_cases != 1
    }
    \item P - number of phenotypes
    \item sample.overlap - sample overlap matrix
    \item sum.stats - processed summary statistics (SNP aligned effect sizes, subsetted to common SNPs across data sets, effect sizes converted to Z, etc)
    \item ref.prefix - genotype reference data prefix
    \item analysis.snps - subset of SNPs shared across all data sets
    \item unalignable.snps - SNPs removed during alignment
    \item ref - environment containing the genotype reference data bim file (ref$bim)
}
}
\description{
Processes summary statistics and related information (e.g. sample overlap, case/control ratios).
Ensures alignment of SNPs across data sets
}
