% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{process.locus}
\alias{process.locus}
\title{Process locus info}
\usage{
process.locus(locus, input, min.K = 2, prune.thresh = 99)
}
\arguments{
\item{input}{Input object created with the \code{\link{process.input}} function, containing relevant summary statistics and related info (e.g. sample overlap, case/control ratio)}

\item{min.K}{Minimum number of PCs required to process locus (cannot be less than two). If this criterion is not met, the function will return NA and the locus cannot be analysed.}

\item{prune.thresh}{PC pruning threshold governing the maximum number of PCs to retain.
Selects PCs as such that the cumulative proportion of variance explained is at least that of the threshold (set to 99 percent by default).}

\item{loc}{Locus info for a single locus, obtained using the \code{\link{read.loci}} function. Expects a locus ID ('LOC') together with locus coordinates ('CHR', 'START', 'STOP') and/or a ';' separated SNP list ('SNPS')}
}
\value{
This function returns an object containing general locus info, the relevant processed sumstats, and all info about the input phenotypes required for analysis

\itemize{
    \item id - locus ID
    \item chr/start/stop - locus coordinates
    \item snps - list of locus SNPs
    \item n.snps - number of SNPs
    \item K - number of PCs
    \item delta - PC projected joint SNP effects
    \item sigma - sampling covariance matrix
    \item omega - genetic covariance matrix
    \item omega.cor - genetic correlation matrix
    \item N - vector of average N across locus SNPs for each phenotype
    \item phenos - phenotype IDs
    \item binary - boolean vector indicating whether phentoypes are binary
}
}
\description{
Takes a single locus and computes all the relevant parameters for this locus using the provided summary statistics and related info.
}
