% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/constrain.R
\name{Range.lvm}
\alias{Range.lvm}
\title{Define range constraints of parameters}
\usage{
Range.lvm(a = 0, b = 1)
}
\arguments{
\item{a}{Lower bound}

\item{b}{Upper bound}
}
\value{
function
}
\description{
Define range constraints of parameters
}
\author{
Klaus K. Holst
}

