% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{revdiag}
\alias{revdiag}
\alias{revdiag<-}
\title{Create/extract 'reverse'-diagonal matrix}
\usage{
revdiag(x,...)

revdiag(x,...) <- value
}
\arguments{
\item{x}{vector}

\item{value}{For the assignment function the values to put in the diagonal}

\item{\dots}{additional arguments to lower level functions}
}
\description{
Create/extract 'reverse'-diagonal matrix
}
\author{
Klaus K. Holst
}

