% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run.multireg}
\alias{run.multireg}
\title{Multiple local genetic regression analysis}
\usage{
run.multireg(
  locus,
  phenos = NULL,
  adap.thresh = c(1e-04, 1e-06),
  only.full.model = F,
  p.values = T,
  CIs = T,
  param.lim = 1.5,
  suppress.message = F
)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function, containing all the relevant parameters for the phenotypes of interest}

\item{phenos}{Optional parameter specifying subset of phenotypes. If NULL, all phenotypes in the locus object will be analysed. NOTE: the last phenotype will be treated as the outcome}

\item{adap.thresh}{Adaptive iteration procedure for permutation p-values that increases the number of permutations for lower p-values. Can be set turned of by setting to NULL.}

\item{p.values}{Set to F if p-values are not desired.}

\item{CIs}{Set to F to if confidence intervals are not desired.}

\item{param.lim}{The +- threshold at which estimated parameters are considered to be too far out of bounds (and will be set to NA)}
}
\description{
Multiple local genetic regression analysis
}
