% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run.partial.cor}
\alias{run.partial.cor}
\title{Partial correlation analysis}
\usage{
run.partial.cor(
  locus,
  phenos = NULL,
  adap.thresh = c(1e-04, 1e-06),
  CIs = T,
  p.values = T,
  max.r2 = 0.95,
  param.lim = 1.25
)
}
\arguments{
\item{locus}{Locus object}

\item{phenos}{List of phenotypes to analyse (first two are p1 & p2, the remaining are Z)}

\item{CIs}{Set to F to if confidence intervals are not desired.}

\item{p.values}{Set to F if p-values are not desired.}

\item{max.r2}{Max r2 threshold for the regression of x~z and y~z. If any of these r2's are too high, the partial correlation becomes unstable, and analysis is therefore aborted.}

\item{adapt.thresh}{Adaptive iteration procedure for permutation p-values that increases the number of permutations for lower p-values. Can be set turned of by setting to NULL.}
}
\description{
Will perform the partial correlation between the first two phenotypes (p1,p2) conditioned on the rest (Z). 
Phenotype order is based on that within the locus object by default, but can be changed by passing a phenotype vector to the 'phenos' argument.
}
