% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run.univ.bivar}
\alias{run.univ.bivar}
\title{Perform univariate and bivariate analysis, with univariate filtering}
\usage{
run.univ.bivar(
  locus,
  phenos = NULL,
  univ.thresh = 0.05,
  adap.thresh = c(1e-04, 1e-06),
  return.unanalysed = F
)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function}

\item{phenos}{Optional parameter specifying subset of phenotypes. If NULL, all phenotypes will be analysed}

\item{univ.thresh}{P-value threshold for the univariate test used to determine whether a bivariate test should be performed.}

\item{adap.thresh}{P-value thresholds for the adaptive permutation procedure, determining the number of iterations to use for the bivariate test.
Default number of iterations is 1e+4, but will be increased to 1e+5, and 1e+6 as p-values fall below the respective thresholds.
If set to NULL, the maximum number of permutations is capped at the default (note: this significantly speeds up the analysis, but results in poor accuracy for low p-values).}

\item{return.unanalysed}{If true, the bivariate results will contain all phenotypes, with NAs for those that weren't analysed.
Otherwise, only the analysed phenotypes will be returned (default).}
}
\value{
List containing the results from the univariate and bivariate tests
}
\description{
Will only perform bivariate test for phenotypes that pass the univariate significance threshold.
If multiple phenotypes are specified, the last will be treated as the target phenotype, and the bivariate test will be performed between this phenotype and all others
}
