% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run.univ}
\alias{run.univ}
\title{Run univariate analysis}
\usage{
run.univ(locus, phenos = NULL)
}
\arguments{
\item{locus}{Locus object created using the the \code{\link{process.locus}} function}

\item{phenos}{Optional parameter specifying subset of phenotypes. If NULL, all phenotypes will be analysed}
}
\description{
Performs univariate test to determine the presence of local genetic signal (i.e. local heritability)
}
