% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{serotonin}
\alias{serotonin}
\title{Serotonin data}
\format{data.frame}
\source{
Simulated
}
\description{
This simulated data mimics a PET imaging study where the 5-HT2A
receptor and serotonin transporter (SERT) binding potential has
been quantified into 8 different regions. The 5-HT2A
cortical regions are considered high-binding regions
measurements.  These measurements can be regarded as proxy measures of
the extra-cellular levels of serotonin in the brain
\tabular{rll}{
        day    \tab numeric \tab Scan day of the year \cr
        age    \tab numeric \tab Age at baseline scan \cr
        mem    \tab numeric \tab Memory performance score \cr
        depr   \tab numeric \tab Depression (mild) status 500 days after baseline \cr
        gene1  \tab numeric \tab Gene marker 1 (HTR2A) \cr
        gene2  \tab numeric \tab Gene marker 2 (HTTTLPR) \cr
        cau \tab numeric \tab SERT binding, Caudate Nucleus \cr
        th  \tab numeric \tab SERT binding, Thalamus \cr
        put \tab numeric \tab SERT binding, Putamen \cr
        mid \tab numeric \tab SERT binding, Midbrain \cr
        aci \tab numeric \tab 5-HT2A binding, Anterior cingulate gyrus \cr
        pci  \tab numeric \tab 5-HT2A binding, Posterior cingulate gyrus \cr
        sfc \tab numeric \tab 5-HT2A binding, Superior frontal cortex \cr
        par \tab numeric \tab 5-HT2A binding, Parietal cortex \cr
}
}
\keyword{datasets}

