% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.default.R
\name{sim.default}
\alias{sim.default}
\alias{as.sim}
\title{Wrapper function for mclapply}
\usage{
\method{sim}{default}(
  x = NULL,
  R = 100,
  f = NULL,
  colnames = NULL,
  messages = lava.options()$messages,
  seed = NULL,
  args = list(),
  iter = FALSE,
  ...
)
}
\arguments{
\item{x}{function or 'sim' object}

\item{R}{Number of replications or data.frame with parameters}

\item{f}{Optional function (i.e., if x is a matrix)}

\item{colnames}{Optional column names}

\item{messages}{Messages}

\item{seed}{(optional) Seed (needed with cl=TRUE)}

\item{args}{(optional) list of named arguments passed to (mc)mapply}

\item{iter}{If TRUE the iteration number is passed as first argument to (mc)mapply}

\item{...}{Additional arguments to (mc)mapply}
}
\description{
Wrapper function for mclapply
}
\seealso{
summary.sim plot.sim print.sim
}
