% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.default}
\alias{sim.default}
\title{Wrapper function for mclapply}
\usage{
\method{sim}{default}(x, R = 100, colnames = NULL, messages = 1L,
  mc.cores = parallel::detectCores(), blocksize = 2L * mc.cores, ...)
}
\arguments{
\item{x}{function or 'sim' object}

\item{R}{Number of replications}

\item{colnames}{Optional column names}

\item{messages}{Messages}

\item{mc.cores}{Number of cores to use}

\item{blocksize}{Split computations in blocks}

\item{...}{Additional arguments to mclapply}
}
\description{
Wrapper function for mclapply
}
\examples{
m <- lvm(y~x+e)
distribution(m,~y) <- 0
distribution(m,~x) <- uniform.lvm(a=-1.1,b=1.1)
transform(m,e~x) <- function(x) (1*x^4)*rnorm(length(x),sd=1)

onerun <- function(iter=NULL,...,b0=1,idx=2) {
    d <- sim(m,2e3,p=c("y~x"=b0))
    l <- lm(y~x,d)
    res <- c(coef(summary(l))[idx,1:2],
             confint(l)[idx,],
             estimate(l,only.coef=TRUE)[idx,2:3:4])
    names(res) <- c("Estimate","Model.se","Model.lo","Model.hi",
                    "Sandwich.se","Sandwich.lo","Sandwich.hi")
    res
}

val <- sim(onerun,R=10,b0=1,messages=0)
val
val <- sim(val,R=100,b0=1) ## append results

summary(val,estimate=c(1,1),confint=c(3,4,6,7),true=c(1,1))
summary(val,estimate=c(1,1),se=c(2,5),names=c("Model","Sandwich"))

if (interactive()) {
    plot(val,1,1)
}
}

