% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack.estimate}
\alias{stack.estimate}
\title{Stack estimating equations}
\usage{
\method{stack}{estimate}(x, model2, D1u, inv.D2u, propensity, dpropensity, U,
  k = 1, keep1 = FALSE, propensity.arg, estimate.arg, ...)
}
\arguments{
\item{x}{Model 1}

\item{model2}{Model 2}

\item{D1u}{Derivative of score of model 2 w.r.t. parameter vector of model 1}

\item{inv.D2u}{Inverse of deri}

\item{propensity}{propensity score (vector or function)}

\item{dpropensity}{derivative of propensity score wrt parameters of model 1}

\item{U}{Optional score function (model 2) as function of all parameters}

\item{k}{Debug argument}

\item{keep1}{If FALSE only parameters of model 2 is returned}

\item{propensity.arg}{Arguments to propensity function}

\item{estimate.arg}{Arguments to 'estimate'}

\item{...}{Additional arguments to lower level functions}
}
\description{
Stack estimating equations
}
