% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stack.estimate}
\alias{stack.estimate}
\title{Stack estimating equations}
\usage{
\method{stack}{estimate}(x, model2, D1u, inv.D2u, weight, dweight, U, k = 1,
  keep1 = FALSE, ...)
}
\arguments{
\item{x}{Model 1}

\item{model2}{Model 2}

\item{D1u}{Derivative of score of model 2 w.r.t. parameter vector of model 1}

\item{inv.D2u}{Inverse of deri}

\item{weight}{weight (vector or function)}

\item{dweight}{derivative of weight wrt parameters of model 1}

\item{U}{Optional score function (model 2) as function of all parameters}

\item{k}{Debug argument}

\item{keep1}{If TRUE only parameters of model 2 i s returned}

\item{...}{Additional arguments to lower level functions}
}
\description{
Stack estimating equations
}
\examples{
2+2
}

