\name{tr}
\alias{tr}
\title{Trace operator}
\usage{
  tr(A)
}
\arguments{
  \item{A}{Square numeric matrix}
}
\value{
  \code{numeric}
}
\description{
  Calculates the trace of a square matrix.
}
\examples{
tr(diag(1:5))
}
\author{
  Klaus K. Holst
}
\seealso{
  \code{\link{crossprod}}, \code{\link{tcrossprod}}
}
\keyword{algebra}
\keyword{math}

