\name{trim}
\alias{trim}
\title{Trim tring of (leading/trailing/all) white spaces}
\usage{
trim(x, all = FALSE, ...)
}
\arguments{
  \item{x}{String}

  \item{all}{Trim all whitespaces?##'}

  \item{\dots}{additional arguments to lower level
  functions}
}
\description{
Trim tring of (leading/trailing/all) white spaces
}
\author{
Klaus K. Holst
}

