% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.error.R
\name{twostage}
\alias{twostage}
\title{Two-stage estimator}
\usage{
twostage(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{Additional arguments to lower level functions}
}
\description{
Generic function
}
\seealso{
twostage.lvm twostage.lvmfit twostage.lvm.mixture twostage.estimate
}

