% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{twostageCV}
\alias{twostageCV}
\title{Cross-validated two-stage estimator}
\usage{
twostageCV(model1, model2, data, control1 = list(trace = 0),
  control2 = list(trace = 0), knots.boundary, mc.cores = 1, K = 1:4,
  nknots = 1:9, fix = TRUE, std.err = TRUE, ...)
}
\arguments{
\item{model1}{model 1 (exposure measurement error model)}

\item{model2}{model 2}

\item{data}{data.frame}

\item{control1}{optimization parameters for model 1}

\item{control2}{optimization parameters for model 1}

\item{knots.boundary}{boundary points for natural cubic spline basis}

\item{mc.cores}{number of cores to use for parallel computations}

\item{K}{number of mixture components}

\item{nknots}{number of knots}

\item{fix}{automatically fix parameters for identification (TRUE)}

\item{std.err}{calculation of standard errors (TRUE)}

\item{...}{additional arguments to lower level functions}
}
\description{
Cross-validated two-stage estimator for non-linear SEM
}
